requirements.txt
pandas
numpy
scikit-learn
matplotlib
librosa
tensorflow
torch
brainflow
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqtgraph==0.12.4
pyyaml 